refactor: move driver repository implementation of firestore and in-memory to adapter/gateways/database folder.
